pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "saga"
version = "1.0.0"
description = "saga"
requires-python = ">=3.8"
readme = {file = "README.md", content-type = "text/markdown"}
dependencies = [
    "fourm @ git+https://github.com/apple/ml-4m.git@main#egg=fourm",
    "omegaconf",
    "hydra-core",
    "mediapy",
    "loguru",
    "ipykernel",
    "numpy",
    "pandas",
    "opencv-python",
    "pydub",
    "yt-dlp",
    "torch",
    "torchvision",
    "torchaudio",
    "transformers",
    "pillow",
    "torchaudio",
    "jsonlines",
    "pytube",
    "notebook"
]

[project.urls]

[tool.setuptools.packages.find]
exclude = ["cfgs*", "assets*", "notebooks*"]

[tool.wheel]
exclude = ["cfgs*", "assets*", "notebooks*"]
